feature: 로그아웃 폼 외부 클릭, 페이지 이동 시 자동으로 닫히도록 구현
- 로그아웃 폼 연 상태에서 다른 모달 창 열었을 때 안닫히는 버그 존재

resolved:#96
